requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlmodel>=0.0.8
psycopg2-binary>=2.9.9
alembic>=1.12.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
email-validator>=2.0.0
celery>=5.2.7
redis>=4.5.1
kafka-python>=2.0.2
